requirements.txt
# 中国股市 AI 分析程序依赖
# Chinese Stock Market AI Analysis Program Dependencies

# ==================== Core Framework ====================
# Web 框架 / Web Framework
streamlit>=1.28.0
plotly>=5.18.0

# ==================== Quantitative Analysis ====================
# qlib 量化框架 / qlib Quant Framework
pyqlib>=0.9.0

# ==================== Data Processing ====================
# 数据处理库 / Data Processing Libraries
pandas>=2.0.0
numpy>=1.22.0,<2.1.0  # 兼容 gensim 和 numba
scipy>=1.11.0

# ==================== Data Sources ====================
# 中国股市数据源 / Chinese Stock Market Data Sources
tushare>=1.3.0
akshare>=1.11.0

# ==================== Technical Analysis ====================
# 技术分析库 / Technical Analysis Library
# 注意: TA-Lib 需要先安装系统依赖 / Note: TA-Lib requires system dependencies first
# macOS: brew install ta-lib
# Ubuntu: sudo apt-get install ta-lib
TA-Lib>=0.4.28

# ==================== Machine Learning ====================
# 机器学习库 / Machine Learning Libraries
lightgbm>=4.1.0
scikit-learn>=1.3.0
torch>=2.0.0

# ==================== Data Storage ====================
# 数据存储 / Data Storage
sqlalchemy>=2.0.0
pymysql>=1.1.0

# ==================== Utilities ====================
# 工具库 / Utility Libraries
python-dotenv>=1.0.0
loguru>=0.7.0
pyyaml>=6.0
requests>=2.31.0

# ==================== Optimization ====================
# 组合优化库 / Portfolio Optimization
cvxpy>=1.4.0
pyportfolioopt>=1.5.0

# ==================== Testing ====================
# 测试框架 / Testing Framework
pytest>=7.4.0
pytest-cov>=4.1.0